cleanup

+++ client/src/redux/socket/socketIO.js
@@ -3,6 +3,9 @@ import { getNewRoomData } from "../features/rooms/roomSlice";
 import { addRoomToUserRoomsList, socketIoLeaveRoom, removeRoomFromUserSlice } from "../features/users/userSlice";
 import { removeRoomFromRoomSlice } from "../features/rooms/roomSlice";
 import { pushMessageToChatLog, getNewChatLogData, socketIoJoinRooms, removeRoomFromChatLogSlice } from "../features/chatLogs/chatLogSlice";
+import { updateActivityLog } from "../features/activityLogs/activityLogSlice";
+import { setLastActiveInLocalStorage } from "../../functions/misc/localStorage";
+import { updateAlertLog } from "../features/globalAlerts/globalAlertSlice";
 
 
 
@@ -20,7 +23,6 @@ export const socket = io(URL, {
 export const socketIoListenForMessage = (dispatch) => {
     socket.on("message", (messageData) => {
         dispatch(pushMessageToChatLog(messageData))
-        console.log("new message", messageData)
     })
 }
 
@@ -30,34 +32,83 @@ export const socketIoHeartbeat = () => {
     })
 }
 
-export const socketIoListenForGlobalAlert = (dispatch, userId, userRooms) => {
+export const socketIoListenForGlobalAlert = (dispatch, userId, userRooms, activityLog) => {
     socket.on('globalAlert', (data) => {
-
-        console.log("global alert", data)
-        console.log("userid", userId)
-        console.log(data.eventType)
 
         if (data.eventType === 'newRoomCreated') {
             const newRoomId = data.eventData.roomId
-            if (data.eventData.users.includes(userId) ) {
+            // if (data.eventData.users.includes(userId) ) {
+          
+            // }
+            if (data.eventData.users.includes(userId) && (data.eventData.createdBy !== userId)) {
                 dispatch(socketIoJoinRooms([newRoomId]))
-            }
-            if (data.eventData.users.includes(userId) && data.eventData.createdBy !== userId) {
+                
                 dispatch(addRoomToUserRoomsList(newRoomId))
                 dispatch(getNewRoomData(newRoomId))
                 dispatch(getNewChatLogData(newRoomId))
+
+                // try this instead to prevent double render
+                // getUserByMongoBbUserId
+                // getRoomsByMongoDbUserId
+
+                // update redux activity log, roomId: time=now
+                const date = Date.now()
+                dispatch(updateActivityLog({
+                    ...activityLog,
+                    roomId: newRoomId,
+                    timestamp: date
+                }))
+
+                setLastActiveInLocalStorage(userId, {
+                    ...activityLog,
+                    [newRoomId]: date
+                })
+
+                dispatch(updateAlertLog({
+                    alertTitle: "New Group Created",
+                    alertTime: date,
+                    alertInfo: {
+                        roomName: data.eventData.roomName,
+                        createdByUsername: data.eventData.createdByUsername
+                    }
+                }))
             }
+
         }
 
         if (data.eventType === 'usersAddedToRoom') {
             const addedUserIdsArr = data.eventData.addedUsers
-            if (addedUserIdsArr.includes(userId)) {
+            const newRoomId = data.eventData.roomId
+            if (addedUserIdsArr.includes(userId) && (userId !== data.eventData.updatedById)) {
                 const roomId = data.eventData.roomId
                 dispatch(socketIoJoinRooms([roomId]))
 
                 dispatch(addRoomToUserRoomsList(roomId))
                 dispatch(getNewRoomData(roomId))
                 dispatch(getNewChatLogData(roomId))
+
+                // update redux activity log, roomId: time=now
+                const date = Date.now()
+                dispatch(updateActivityLog({
+                    ...activityLog,
+                    roomId: newRoomId,
+                    timestamp: date
+                }))
+
+                setLastActiveInLocalStorage(userId, {
+                    ...activityLog,
+                    [newRoomId]:  date
+                })
+
+                dispatch(updateAlertLog({
+                    alertTitle: "User Added To Group",
+                    alertTime: date,
+                    alertInfo: {
+                        roomName: data.eventData.roomName,
+                        updatedByUsername: data.eventData.updatedByUsername
+                    }
+                }))
+
             }
         }
 
@@ -67,7 +118,7 @@ export const socketIoListenForGlobalAlert = (dispatch, userId, userRooms) => {
             // so no updates
             // add to removed rooms list
             const roomId = data.eventData.roomId
-            if (data.eventData.users.includes(userId)) {
+            if (data.eventData.addedUsers.includes(userId) && (userId !== data.eventData.updatedById)) {
                 dispatch(socketIoLeaveRoom(roomId))
 
                 // update rooms, user, chatLog slices
@@ -75,7 +126,18 @@ export const socketIoListenForGlobalAlert = (dispatch, userId, userRooms) => {
                 dispatch(removeRoomFromRoomSlice(roomId))
                 dispatch(removeRoomFromChatLogSlice(roomId))
 
+                // also rerender info for all users in the group
+
                 // add alert / notification
+                const date = Date.now()
+                dispatch(updateAlertLog({
+                    alertTitle: "User Removed From Group",
+                    alertTime: date,
+                    alertInfo: {
+                        roomName: data.eventData.roomName,
+                        updatedByUsername: data.eventData.updatedByUsername
+                    }
+                }))
             }
         }
 
@@ -85,8 +147,7 @@ export const socketIoListenForGlobalAlert = (dispatch, userId, userRooms) => {
             // so no updates
             // add to deleted rooms list
             const roomId = data.eventData.roomId
-            console.log("room DELETED", userRooms)
-            if (userRooms?.includes(roomId)) {
+            if (userRooms?.includes(roomId) && userId !== data.eventData.deletedBy) {
                 dispatch(socketIoLeaveRoom(roomId))
 
                 dispatch(removeRoomFromUserSlice(roomId))
@@ -96,6 +157,36 @@ export const socketIoListenForGlobalAlert = (dispatch, userId, userRooms) => {
                 // update rooms, user, chatLog slices
 
                 // add alert / notification
+                const date = Date.now()
+                dispatch(updateAlertLog({
+                    alertTitle: "Group Deleted",
+                    alertTime: date,
+                    alertInfo: {
+                        roomName: data.eventData.roomName,
+                        deletedBy: data.eventData.deletedByUsername
+                    }
+                }))
+            }
+        }
+
+        if (data.eventType === 'roomNameUpdated') {
+            const roomId = data.eventData.roomId
+            if (userRooms?.includes(roomId) && userId !== data.eventData.updatedById) {
+                // dispatch(socketIoLeaveRoom(roomId))
+                // update rooms, user, chatLog slices
+
+                // add alert / notification
+                const date = Date.now()
+                dispatch(updateAlertLog({
+                    alertTitle: "Group Name Updated",
+                    alertTime: date,
+                    alertInfo: {
+                        originalRoomName: data.eventData.originalRoomName,
+                        roomName: data.eventData.roomName,
+                        updatedByUsername: data.eventData.updatedByUsername,
+                    }
+                }))
+
             }
         }
